# Move assertion out of loop for better performance